Add support for decorator attributes in StyledBox
